Major bug fix. There was a problem in the 'Dao.Text.Regex.regexToTable' and 'Dao.Text.Regex.checkRegexChoices' functions, and the instantiation of 'Dao.Text.Regex.RegexTable' into the 'Data.Monoid.Monoid' type class. All of these functions assumed that when taking a list of regular expressions to be joined in parallel, all of the items in the first list must be checked against all of the items in the second list to see if any regular expressions being shadowed.
Before this commit, these functions were all only checking the last
element of the left list with the first element of the right list to
see if those two expressions were shadowing. This will not catch all
cases of shadowing, as I discovered the hard way during testing.

My update has been tested and it works as it is supposed to, for the
most part. Still I have not been able to join equivalent regular
expressions within table elements, so if there are two equivalent
regular expressions, both will still be tried in turn. This will need
to be fixed in another update.