Did a rather substantial rewrite of the machinery of the 'Dao.Rule.Rule' class. On the surface, only two API functions have changed. (1) The 'limitByScore' function has been removed, and it may be restored in the near future but for now there is no good way to implement it given the changes. (2) The 'guessPartial' function has been added, which makes it easier to implement predictive input (e.g. tab completion on command lines).
The biggest change I have made is that the 'Dao.Rule.StaefulRule' monad
no longer lifts the 'Dao.Logic.LogicT' monad, and as a result
'StatefulRule' no longer instantiates 'Dao.Logic.MonadLogic'. Now,
'StatefulRule' could be said to "compile" to a 'LogicT' monad by way of
the 'evalRuleLogic' function, but 'StatefulRule' is otherwise a
completely independent domain specific language now.

The reason this was necessary was because I had discovered that partial
queries performed by the 'Dao.Rule.partialQuery' function were
impossible when an arbitrary 'Dao.Logic.LogicT' monad could be lifted.
The reason this was impossible was because some functions, for example
the 'Dao.Rule.bestMatch' function, were able to collapse a
'Dao.Rule.StatefulRule' using the 'Dao.Rule.evalRuleLogic' arbitrarily.
This meant that using the 'bestMatch' function would result in the
'StatefulRule' being collapsed into a 'LogicT' monad, then lifted back
into a 'StatefulRule' after some modifications to the state had been
made. When this happened, all information about the rule tree was
evaluated to a 'LogicT' monad and therefore made inaccessible to the
'partialQuery' function.

The 'Dao.Rule.StatefulRule' monad needs to be homogeneous monadic data
structure, like a list monad. Functions must not have the ability to
obscure the data structure by evaluating it to an opaque function data
type and re-lifting the evaluated bit back into the monad. This was the
change I made; it is now impossible to lift a 'LogicT' monad into a
'Dao.Rule.StatefulRule', and the tree structure is therefore always
available for analysis. The 'Dao.Logic.LogicT' monad is still useful as
an intermediate step during evaluation of the rule to it's lifted monad.