This commit publishes a major rewrite to the "Dao.Object" module. I have updated the 'Dao.Object.Simple' data type such that it may no longer contain 'Dao.Object.Object' data types, only other 'Simple' data types. As a result, 'Simple' is now something of a protocol for transmitting and receiving data types.
Similarly 'Object' no longer contains the 'Simple' data type unless a
'Simple' type is converted to an 'Object' by way of the
'Dao.Object.fromForeign' or 'Dao.Object.obj' functions. 'Object' is
now essentially a superset of the 'Data.Dynamic.Dynamic' data type, as
in it is like 'Data.Dynamic.Dynamic' but has more features, for
example the ability to be converted to a 'Simple' type or the ability
to be 'pPrint'-ed without unwrapping it.

However, the 'Dao.Object.SimpleData' class is still a superclass of
'Dao.Object.ObjectData', so you cannot instantiate a data type into
the 'ObjectData' class without instantiating the 'SimpleData' class.

Also, the 'Simple' data type has been extended with the
'Dao.Object.OStruct' constructor, which operates kind of like
Haskell's "newtype" keyword. It provides a labeled wrapper for an
existing 'Simple' type.

Finally, I had to rename the 'Dao.Rule.tree' and 'Dao.Rule.struct'
functions, because the name 'struct' is now taken by the "Dao.Object"
module, and also I thought the name 'tree' was misleading.