Made the 'Dao.Rule.Rule' data type extremely complicated and nearly impossible to understand, but unfortunately, it was necessary.
I really don't like data types which are polymorphic over more than 3
variables. 'Dao.Rule.Rule' was at my limit, it a monad transformer, so
already it is polymorphic over the return type and the lifted monadic
type. Then it allowed a polymorphic error type, and that was my limit.
I was squeamish to add the polymorphic stateful type, but there is
just no better way backtrack updates to state as rules are evaluated.

But now I have really gone and done it: I have made the token type
polymorphic (where it used to be 'Dao.Object.Object'), and I have made
the pattern type polymorphic (which also used to be a
'Dao.Object.Object'). So now 'Dao.Rule.Rule' is polymorphic over six
types.

But using duck typing for patterns and tokens in 'Dao.Rule.Rule's was
just too limited, it was too easy to confuse patterns with tokens, and
it was too difficult to define functions that can define matching
behavior for patterns.

Now the 'queryAll' and related functions all take a function of type
'Dao.Rule.MatchFunction', which can be monadic, which defines how to
match a token to a pattern. This is more appropriate than requiring
your pattern type to install a matching method when instantiating
'Dao.Object.ObjectData'.

All code has been tested, and the modifications are working properly.