Did a total rewrite of everything. There are number similarities between Dao-2.0 and Dao-1.0, and it would have been better to make these modifications piecemeal and commit and document each change separately, but I am too short on free time, and anyway nobody cares what I do here, this is my own project.
The biggest change is that Dao-2.0 is a Lisp-like *protocol*, and rather than
whole programming language. So Dao Lisp is a language intended to be written and
read by computers, not humans (that is the difference between a protocol and a
programming language). As a protocol, the Dao parser has been greatly simplified
to a minimalistic 'readsPrec' parser.

The Dao-1.0 'Object' primitive type has been changed to a 'DaoExpr' in Dao-2.0
type but is mostly similar to Dao-1.0. Binary serialization has not been
implemented yet, but it will be soon. The module hierarchy has been changed from
it's own top-level "Dao" to "Language.Interpreter.Dao", and the ENTIRE
interpreter and all primitives have been moved into a single "Kernel" module
which defines the core of the Dao protocol, syntax, and runtime semantics.

Everything else has been coldly deleted.